Update pin_configurations.dart, pin_entry.dart, pin_input_field.dart, and pin_keyboard.dart.
